Only import psutil when the library is installed
Since we don't like to force pyctuator users to use the psutil library,
we'll import it and use it only if its available.

To test this, manually delete the library from pipenv and verify the API
is working and that the memory/thread-count metrics are not shown and
that no error is raised.