Added disk-space monitoring to health endpoint
While the health endpoint need to be generalized and extended, the the
disk-space health was added as a "provider" so other health checks could
be made providers.

The status of the entire service (the topmpost status in the health
endpoint) is naively determined by logical-and of all health statuses
when "UP" is regarded as "True".

Since the disk-space data is retrieved using the opional psutil
libraray, the code will access psutil only if it is installed.