Don't use global variables for the boot-admin registration
In order for the recurring/scheduled registration to be stoppable (by
tests for example), it used to maintain a global variable telling it its
ok to continue.

This is a bad practice (google "static is evil").

Instead, the registration loop is now encapsulated in a class which
maintains its state in member variables.
Not only getting rid of the static flags, this simplified and reduced
the code a lot.