Generalize health status using a list of health providers
Pyctuator derives the application's health status by combining the
health status returned from a list of health-providers.
A health-provider provides its name, and current status and health.
The overall health status will be "UP" if the status of all providers
was "UP", in any other case, the overall status will be "DOWN".