Verify Pyctuator behaves when psutil is not installed
Added a test that verifies the behavior of Pyctuator when psutil library
isn't installed - the tests verifies that the diskHealth section in the
health response is missing.

To do so, all non-dev dependencies were made optionals so "poetry
install" doesn't install them. A test was added that is running only of
psutil is not installed, the test asserts that "diskSpace" is missing
from the health json response.
All other tests that assume psutil is installed, are skipped if psutil
isn't installed.

The .gitlab-ci.yml was modified to first tests without psutil and then
with psutil.