Override Flask's jsonify()'s default datetime format to use iso.
By default, when Flask is rendering a response to JSON, it is formatting
datetime, date and time according to RFC 822 which is different from the
ISO format used by SBA.

With this fix, we now call `jsonify()` directly instead of converting to
`Dict`, which also takes care for setting the response's content-type.
See `flask.json.JSONEncoder`.

Before this fix, SBA didn't display the http-traces plots because the
datetime in every request-response pair wasn't formatted properly.