Mark pyctuator as a type-hinted package
Tools like mypy need a way to know that the package contains type-hints.
According to PEP-561, the way to do it in a package that contains
type-hints inline is adding a py.typed file in the package root.

Tested by installing the package and running mypy, seems to work well.