Examples, http-trace allow "Exclude /pyctuator/**", dont request logging
In order for the "Http Traces" tab to be able to hide requests sent by
Spring Boot Admin to the Pyctuator endpoint, `pyctuator_endpoint_url`
must be using the same host and port as `app_url`.
The examples initially registered the host's IP for `app_url` instead
of using the same host as `pyctuator_endpoint_url` - the idea was that
the app's URL shown in SBA will be accessible.

Also, both FastAPI and Flask examples were logging every request sent
from SBA to Pyctuator. This was annoying so the example now are
configuring the loggers not to show access-log.