Http-traces's `timeTaken` should be millis between request and response
When http-traces is capturing a request and response, the `timeTaken`
field should be millis between the time the request was received and the
time the response was sent.

Both `fastapi_pyctuator` and `flask_pyctuator` used the `microseconds`
of the `timedelta` between the request time and the response time.
Unfortunately, this only returns the microseconds part of the time
delta. Instead, should calculate the difference between the two
timestamps.

Fix #13