Tornado HTTP traces should support multi-value headers
As reported in #46, `TornadoHttpPyctuator` doesn't properly handle
multi-value HTTP headers, which are supproted by
`tornado.httputil.HTTPHeaders` as multiple entries with the same header.

More than that, the value put in traces was a string instead of a
list-of-strings.

Note that `requests` doesn't seem to support sending a request with
multi-value headers, so the testing is limited to verifying
`TornadoHttpPyctuator._get_headers` only.